Error handling

+++ src/main/java/com/mycompany/services/EmployeeService.java
@@ -13,7 +13,13 @@ public class EmployeeService implements AbstractEmployeeService {
         employeeDAO = new EmployeeDAO(props);
     }
 
-    public boolean createEmployee(Employee employee){
+    public boolean createEmployee(Employee employee) throws Exception {
+        if(employee.getId() == 0){
+            throw new Exception("Id cannot be null");
+        }
+        if(employee.getName() == null || employee.getDesignation()==null){
+            throw new Exception("Name/Designation both cannot be null");
+        }
         return employeeDAO.createEmployee(employee);
     }
 
@@ -21,11 +27,31 @@ public class EmployeeService implements AbstractEmployeeService {
         return employeeDAO.getEmployee(empId);
     }
 
-    public boolean updateEmployee(Employee employee){
+    public boolean updateEmployee(Employee employee) throws Exception {
+        if(employee.getId() == 0){
+            throw new Exception("Id cannot be null");
+        }
+        if(employee.getName() == null || employee.getDesignation()==null){
+            throw new Exception("Name/Designation both cannot be null");
+        }
         return employeeDAO.updateEmployee(employee);
     }
 
-    public boolean deleteEmployee(int empId){
+
+    public boolean patchEmployee(Employee employee) throws Exception {
+        if(employee.getId() == 0){
+            throw new Exception("Id cannot be null");
+        }
+        if(employee.getName() == null && employee.getDesignation()==null){
+            throw new Exception("Name and Designation both cannot be null");
+        }
+        return employeeDAO.updateEmployee(employee);
+    }
+
+    public boolean deleteEmployee(int empId) throws Exception {
+        if(empId == 0){
+            throw new Exception("Id cannot be null");
+        }
         return employeeDAO.deleteEmployee(empId);
     }
